refactor filters and handlers to use InPrivateChat for private message handling

+++ src/app/handlers/callback.py
@@ -1,6 +1,6 @@
 from aiogram import F, Router
 from aiogram.fsm.context import FSMContext
-from aiogram.types import CallbackQuery
+from aiogram.types import CallbackQuery, Message
 
 import app.utils.keyboards as kb
 from app.utils.states import SupportStates
@@ -10,7 +10,10 @@ router = Router()
 
 @router.callback_query(F.data == 'support')
 async def support(call: CallbackQuery, state: FSMContext) -> None:
-    await call.message.answer('Введите ваше сообщение:')  # type: ignore
+    if not call.message:
+        return
+
+    await call.message.answer('Введите ваше сообщение:')
     await state.set_state(SupportStates.get_msg)
 
 
@@ -18,8 +21,8 @@ async def support(call: CallbackQuery, state: FSMContext) -> None:
 async def end_dialog(call: CallbackQuery, state: FSMContext) -> None:
     await state.clear()
 
-    if call.message:
-        await call.message.delete()  # type: ignore
+    if call.message and isinstance(call.message, Message):
+        await call.message.delete()
         await call.message.answer(
             text=(
                 'Спасибо за обращение! Вы всегда можете снова обратиться '
